Split out uncertainty report print out from uncertainty plot methods
First steps to splitting out the uncertainty report print out from the uncertainty plot methods. views.plot_all still includes this code and should be removed in future.

This is required to make a more modular set of save out methods for a universal save_out function for both the GUI and the CLI.