Mark inner loop variables as private for openmp
For nested loops, the inner loop variable needs to be private, else the
parallelised loops will interfere with each other and return early.
OPENMP automatically marks any variables that are declared within the
parallelised block as being private but if the inner loop variable is
declared at the top of the function, then this does not happen.