Compat for both scipy 1.8 and 1.10 in entropy calc
scipy 1.10 removed the .cov accessor to the covariance; there is no
simple code that works with both scipy 1.8 and 1.10 so it is necessary
to test for the presence of the cov attribute prior to use.

Thanks to Drew Parsons for the patch via the Debian packaging for bumps.

See https://github.com/bumps/bumps/issues/118